Tracing and tweaking infrastructure
This adds a fine-grained tracing and tweaking mechanism, which can
be used to log information about specific topics.
(see https://bugs.freedesktop.org/show_bug.cgi?id=91838)

The spice-traces.def file defines the available traces and tweaks.
It defines 'traces' (boolean flags) and 'tweaks' (int values).
Traces are intended to activate specific debug printouts, whereas
tweaks are intended to control specific 'tuning' parameters.

The TRACE(trace_name) and IFTRACE(trace_name) macros can be used to
test the value of a trace flag.

The spice_set_trace function sets traces, e.g. from the command line
or environment variable SPICE_TRACES. The syntax is a colon-separated
list of trace assignments for the form 'name' or 'name=value'. In the
first form, a value of 1 is used. The name can optionally be preceded
by a '*' sign, in which case all traces matching the name will be
affected.

For example,
- SPICE_TRACES='ssl' activates debug printouts related to SSL.
- SPICE_TRACES='ssl:channels' activates debug printouts related to
  either SSL or channels.
- SPICE_TRACES='threads=8' sets the 'threads' tweak to 8.
- SPICE_TRACES='*debug=1' sets all traces containing 'debug' to 1.