common: Connect spice logging to flight recorder
The existing spice macros are connected to the flight recorder,
and the flight recorder output is connected by default to glib logging.
By default, warning, errors and critical messages are "traced", i.e.
displayed as they happen.

The flight recorder output can be configured with SPICE_TRACES or
RECORDER_TRACES, and this can override default settings, e.g. disable
output for critical messages, or change crash dump behavior.

If the RECORDER_TRACES variable is used to configure the recorder,
then output is not connected to glib logging but uses the default
recorder output format.