Add flight recorder and connect default log functions to it
The flight recorder (https://github.com/c3d/recorder) is a non-blocking
logging system with a printf-like interface, which makes it possible to
record event at sufficiently high speed that the instrumentation can be
left active all the time even in production systems.

The flight recorder stores data continuously in multiple circular
buffers called "recorders". Each individual recorder has a purpose
(visible in on-line help) and size (usually between 16 and 128).
Having multiple independent buffers allows important low-freqency
events (e.g. program command-line options) to not be overwritten by
high-frequency events that happen later.

The proposed connection to SPICE through the spice-common existing
logging macros features three main modes of operation:

1. Passive flight recording: In this mode, data is stored continuously
   in circular buffers, and can be replayed e.g. from a debugger
   (e.g. by using "print recorder_dump()" from gdb), or when a signal
   arrives (in which case the captured logs can give hints as to what
   happened before the signal).

2. Active tracing: By setting the SPICE_TRACES or RECORDER_TRACES
   environment variable, it is possible to to put one or several
   individual recorders in "tracing" mode, in which case anything
   recorded is also printed out immediately.

3. Signal dump: When a signal is received, the flight recorder is
   "dumped", i.e. all the content is printed through the standard
   logging mechanism. If a program is locked up, slow or otherwise
   misbehaving, it is possible to get a dump at any time by sending
   a signal such as SIGUSR1 (Linux) or SIGINFO (BSD).

The intent of this patch is simply to provide an initial "connection"
between SPICE and the flight recorder that is as compatible as possible
with the existing logging mechanism (i.e. it still logs through glib
in the default mode of operation). Later patches will provide a
number of useful data probes, notably to enable continuous data
logging of relevant information in the client and server.

This patch also establishes conventions for recorder names in SPICE:

- Recorder names ending in _error, _warning, _critical, _info or _debug
  are given the corresponding GLib log level

- Recorder names ending in _error, _warning or _critical are put in
  tracing mode by default.